requirements.txt
langchain
langgraph
pinecone
openai
google-generativeai
tqdm
requests
python-dotenv 